+++ stat_bsd.go
@@ -9,7 +9,7 @@ import (
 	"time"
 )
 
-const supports supportsType = supportsATime | supportsBTime | supportsCTime | supportsLinks | supportsNice | supportsSymlinks
+const supports supportsType = supportsATime | supportsBTime | supportsCTime | supportsFstat | supportsLinks | supportsNice | supportsSymlinks
 
 const userIDSource UserIDSourceType = UserIDSourceIsInt