+++ stat_supports_unix.go
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: Copyright © 2025 Ross Smith II <ross@smithii.com>
+// SPDX-License-Identifier: MIT
+
+//go:build aix || illumos || solaris
+
+package compat
+
+// Not supported: BTime | Fstat.
+const supports supportsType = supportsLinks | supportsATime | supportsCTime | supportsNice | supportsSymlinks