+++ stat_windows.go
@@ -21,7 +21,7 @@ import (
 
 const perm000 = os.FileMode(0o0)
 
-const supports supportsType = supportsLinks | supportsATime | supportsBTime | supportsCTime | supportsNice | supportsSymlinks
+const supports supportsType = supportsLinks | supportsATime | supportsBTime | supportsCTime | supportsFstat | supportsNice | supportsSymlinks
 
 const userIDSource UserIDSourceType = UserIDSourceIsSID